+++ API/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "node ./packages/api/index.js",
+    "start": "node ./index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -35,7 +35,5 @@
     "openpgp": "^4.10.4",
     "react": "^16.13.1"
   },
-  "devDependencies": {
-    "lerna": "^3.21.0"
-  }
+  "devDependencies": {}
 }
