+++ API/routes/demoReactjs.js
@@ -1,6 +1,6 @@
 var express = require('express');
 var router = express.Router();
-let DB = require('../../../scripts/db');
+let DB = require('../scripts/db');
 
 //handler get home
 router.get('/', function (req, res, next) {
